refactor(views): * split up views into more specific ones * modified load design button * grouped togther 'essentials' * tweaks
